+++ src/hr/avrbanac/openglplayground/entities/Camera.java
@@ -8,7 +8,7 @@ import static org.lwjgl.glfw.GLFW.*;
  * Camera representation class.
  * 
  * @author avrbanac
- * @version 1.0.2
+ * @version 1.0.6
  */
 public class Camera {
     
@@ -16,10 +16,10 @@ public class Camera {
     private float pitch;
     private float yaw;
     private float roll;
-    private long windowID;
+    private long parentWindowID;
     
-    public Camera(long windowID) {
-        this.windowID = windowID;
+    public Camera(long parentWindowID) {
+        this.parentWindowID = parentWindowID;
     }
 
     public Vector3f getPosition() {
@@ -39,16 +39,23 @@ public class Camera {
     }
     
     public void move() {
-        if(glfwGetKey(windowID, GLFW_KEY_W) == GLFW_PRESS) position.z-=0.2f;
-        if(glfwGetKey(windowID, GLFW_KEY_S) == GLFW_PRESS) position.z+=0.2f;
-        if(glfwGetKey(windowID, GLFW_KEY_D) == GLFW_PRESS) position.x+=0.2f;
-        if(glfwGetKey(windowID, GLFW_KEY_A) == GLFW_PRESS) position.x-=0.2f;
-        if(glfwGetKey(windowID, GLFW_KEY_R) == GLFW_PRESS) position.y+=0.2f;
-        if(glfwGetKey(windowID, GLFW_KEY_F) == GLFW_PRESS) position.y-=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_8) == GLFW_PRESS) position.z-=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_2) == GLFW_PRESS) position.z+=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_6) == GLFW_PRESS) position.x+=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_4) == GLFW_PRESS) position.x-=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_7) == GLFW_PRESS) position.y+=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_1) == GLFW_PRESS) position.y-=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_9) == GLFW_PRESS) pitch+=0.2f;
+        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_3) == GLFW_PRESS) pitch-=0.2f;
+        
     }
     
     public void setPosition(Vector3f position) {
         this.position = position;
+    }
+    
+    public void setPitch(float pitch) {
+        this.pitch = pitch;
     }
     
 }